update on logger
