+++ src/evaluate_model.py
@@ -10,8 +10,9 @@ from sklearn.metrics import confusion_matrix
 import pickle
 import sys
 
-logging.basicConfig(level = logging.DEBUG, format = '%(name)-12s %(levelname)-8s %(message)s')
-logger = logging.getLogger('evaluate_model')
+logging.basicConfig(level=logging.DEBUG, filename="logfile", filemode="a+",
+                        format="%(asctime)-15s %(levelname)-8s %(message)s")
+logger = logging.getLogger(__name__)
 
 def evaluate_model(modelpath = None, xtestpath = None, ytestpath = None, **kwargs):
     with open(modelpath, "rb") as f:
