+++ test/test_model.py
@@ -2,6 +2,10 @@ import yaml
 import pandas as pd
 import pickle
 
+logging.basicConfig(level=logging.DEBUG, filename="logfile", filemode="a+",
+                        format="%(asctime)-15s %(levelname)-8s %(message)s")
+logger = logging.getLogger(__name__)
+
 with open('config/config.yml', 'r') as f:
     config = yaml.load(f)
